+++ Modules/Intent.Modules.FastEndpoints/Templates/ApiVersionSwaggerGenOptions/ApiVersionSwaggerGenOptionsTemplatePartial.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Collections.Generic;
+using System.Linq;
 using Intent.Engine;
 using Intent.Modules.Common;
 using Intent.Modules.Common.CSharp.Builder;
@@ -58,6 +59,12 @@ namespace Intent.Modules.FastEndpoints.Templates.ApiVersionSwaggerGenOptions
                         method.AddStatement("return info;");
                     });
                 });
+        }
+
+        public override bool CanRunTemplate()
+        {
+            return ExecutionContext.FindTemplateInstances("Distribution.SwashbuckleConfiguration")?.Any() == true
+                && base.CanRunTemplate();
         }
 
         [IntentManaged(Mode.Fully)]